+++ WatchWise/Controllers/UserWatchedEpisodesController.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 using WatchWise.DTOs.Requests;
 using WatchWise.DTOs.Responses;
 using WatchWise.Services.Implementations;
@@ -19,6 +21,7 @@ namespace WatchWise.Controllers
 
         // GET: api/UserWatchedEpisodes
         [HttpGet]
+        [Authorize(Roles = "ContentManager")]
         public ActionResult<List<UserWatchedEpisodeResponse>> GetUserWatchedEpisodes()
         {
             return Ok(_userWatchedEpisodeService.GetAllUserWatchedEpisodeResponses());
@@ -26,13 +29,19 @@ namespace WatchWise.Controllers
 
         // GET: api/UserWatchedEpisodes/User/5
         [HttpGet("User/{userId}")]
+        [Authorize(Roles = "ContentManager,Subscriber")]
         public ActionResult<List<UserWatchedEpisodeResponse>> GetUserWatchedEpisodesByUserId(long userId)
         {
+            if (User.FindFirstValue(ClaimTypes.NameIdentifier) != userId.ToString() || User.IsInRole("ContentManager") == false)
+            {
+                return Unauthorized();
+            }
             return Ok(_userWatchedEpisodeService.GetUserWatchedEpisodeResponsesByUserId(userId));
         }
 
         // GET: api/UserWatchedEpisodes/Episode/5
         [HttpGet("Episode/{episodeId}")]
+        [Authorize(Roles = "ContentManager")]
         public ActionResult<List<UserWatchedEpisodeResponse>> GetUserWatchedEpisodesByEpisodeId(long episodeId)
         {
             return Ok(_userWatchedEpisodeService.GetUserWatchedEpisodeResponsesByEpisodeId(episodeId));
@@ -40,8 +49,13 @@ namespace WatchWise.Controllers
 
         // POST: api/UserWatchedEpisodes
         [HttpPost]
+        [Authorize(Roles = "Subscriber")]
         public ActionResult PostUserWatchedEpisode(UserWatchedEpisodeRequest userWatchedEpisodeRequest)
         {
+            if (User.FindFirstValue(ClaimTypes.NameIdentifier) != userWatchedEpisodeRequest.UserId.ToString())
+            {
+                return Unauthorized();
+            }
             int addResponse = _userWatchedEpisodeService.AddUserWatchedEpisode(userWatchedEpisodeRequest);
             if (addResponse == -1)
             {
@@ -52,8 +66,13 @@ namespace WatchWise.Controllers
 
         // DELETE: api/UserWatchedEpisodes
         [HttpDelete]
-        public IActionResult DeleteUserWatchedEpisode(UserWatchedEpisodeRequest userWatchedEpisodeRequest)
+        [Authorize(Roles = "Subscriber")]
+        public ActionResult DeleteUserWatchedEpisode(UserWatchedEpisodeRequest userWatchedEpisodeRequest)
         {
+            if (User.FindFirstValue(ClaimTypes.NameIdentifier) != userWatchedEpisodeRequest.UserId.ToString())
+            {
+                return Unauthorized();
+            }
             int deleteResponse = _userWatchedEpisodeService.RemoveUserWatchedEpisode(userWatchedEpisodeRequest);
             if (deleteResponse == -1)
             {
